refactor: Organize window_bounds the same as selector functions
