+++ query/src/func.rs
@@ -1,2 +1,3 @@
 //! Special IOx functions used in DataFusion plans
 pub mod selectors;
+pub mod window;
